+++ src/Skeleton/Uran.java
@@ -24,7 +24,7 @@ public class Uran extends Nyersanyag implements Szallithato{
     	
     	if(expozicio > 2) {
     		a.Robban();
-    		Main.NamesMap.remove(Main.getKeyByValue(Main.NamesMap, this));
+    		Jatek.NamesMap.remove(Jatek.getKeyByValue(Jatek.NamesMap, this));
     	}
     		
     }
